Move kafka connection provider code to a separata class
